requirements.txt
Django==3.2.12
djangorestframework==3.13.1
google-api-python-client==2.38.0
python-decouple==3.6
celery==5.2.3
redis==4.1.4
gunicorn==20.1.0
dj-database-url==0.5.0
psycopg2==2.9.3
whitenoise==6.0.0
django-celery-beat==2.2.1
django-tables2==2.4.1
django-filter==21.1
django-bootstrap3==21.2
